🧹  file structure
