--- ProtocolAddresses.mjs
@@ -1,2 +1,0 @@
-export const LendingPoolConfigurator =
-  '0x949cd89fdcf125e3690235B51D273Ab19bafea7E'
