+++ src/cli/menus/gardenMenu.mjs
@@ -1,0 +1,20 @@
+import prompts from '../prompts/index.mjs'
+import inquirer from 'inquirer'
+import gardenHandlers from '../handlers/gardenHandlers.mjs'
+
+const gardenMenu = async () => {
+  const result = await inquirer.prompt(prompts.menu.garden)
+  let selection
+  switch (result.garden) {
+    case 'addToken':
+      selection = await inquirer.prompt(prompts.garden.addToken)
+      await gardenHandlers.addToken(selection)
+      break
+    case 'executeVote':
+      selection = await inquirer.prompt(prompts.garden.executeVote)
+      await gardenHandlers.executeVote(selection)
+      break
+  }
+}
+
+export default gardenMenu
