+++ src/lib/daoApps.mjs
@@ -1,5 +1,5 @@
 import ethers from 'ethers'
-import { tao_voting } from '../dao.mjs'
+import { tao_voting } from '../config/addresses.mjs'
 import abi from './abis/taoVotingABI.mjs'
 
 // 3. create the voting contract we want to interact with
