+++ src/scripts/lending-pool/activateReserve.mjs
@@ -1,14 +1,12 @@
 import { encodeActCall, encodeCallScript } from '../../lib/evm.mjs'
 import frame from '../../lib/getFrame.mjs'
-import { LendingPoolConfigurator } from '../../ProtocolAddresses.mjs'
-import { tao_agent } from '../../dao.mjs'
+import { LendingPoolConfigurator, tao_agent } from '../../config/addresses.mjs'
 import { TaoVoting } from '../../lib/daoApps.mjs'
 
-// TODO: custom data type
-const signature = ''
+const signature = 'activateReserve(address)'
 const voteDescripton = '0x'
 
-const updateVariableDebtToken = async (args) => {
+const activateReserve = async (args) => {
   const signer = frame()
 
   const lendingPoolCallScript = encodeCallScript([
@@ -35,4 +33,4 @@ const updateVariableDebtToken = async (args) => {
   await votingApp.newVote(agentCallScript, voteDescripton)
 }
 
-export default updateVariableDebtToken
+export default activateReserve
