+++ src/scripts/lending-pool/template.mjs
