+++ src/scripts/lending-pool/updateVariableDebtToken.mjs
@@ -1,13 +1,13 @@
 import { encodeActCall, encodeCallScript } from '../../lib/evm.mjs'
 import frame from '../../lib/getFrame.mjs'
-import { LendingPoolConfigurator } from '../../ProtocolAddresses.mjs'
-import { tao_agent } from '../../dao.mjs'
+import { LendingPoolConfigurator, tao_agent } from '../../config/addresses.mjs'
 import { TaoVoting } from '../../lib/daoApps.mjs'
 
-const signature = 'enableReserveStableRate(address)'
+// TODO: custom data type
+const signature = ''
 const voteDescripton = '0x'
 
-const enableReserveStableRate = async (args) => {
+const updateVariableDebtToken = async (args) => {
   const signer = frame()
 
   const lendingPoolCallScript = encodeCallScript([
@@ -34,4 +34,4 @@ const enableReserveStableRate = async (args) => {
   await votingApp.newVote(agentCallScript, voteDescripton)
 }
 
-export default enableReserveStableRate
+export default updateVariableDebtToken
